[PF-91] Clean file answer_test.js
